Remove conditional code for Python < 3.6.
