+++ setup.py
@@ -19,9 +19,6 @@ if py_version < (3, 6):
 
 packages = ['websockets', 'websockets/extensions']
 
-if py_version >= (3, 6):
-    packages.append('websockets/py36')
-
 ext_modules = [
     setuptools.Extension(
         'websockets.speedups',
